+++ lib/core/constants/app_strings.dart
@@ -1,0 +1,22 @@
+class AppStrings {
+  ///Common
+  static const String appName = 'Clean Architecture App';
+
+  ///Login
+  static const String loginScreenTitle = '$loginText Screen';
+  static const String loginText = 'Login';
+  static const String signupText = 'SIGN UP';
+  static const String nameControllerHintText = 'Enter your Name';
+  static const String nameControllerLabelText = 'Full Name';
+  static const String emailControllerHintText = 'Enter your Email';
+  static const String emailControllerLabelText = 'Email';
+  static const String passwordControllerHintText = 'Enter your Password';
+  static const String passwordControllerLabelText = 'Password';
+
+  ///Home
+  static const String homeScreenTitle = 'Home Screen';
+  static const String welcomeUser = 'Welcome user';
+
+  ///Profile
+  static const String profileScreenTitle = 'Profile Screen';
+}