+++ lib/modules/login/presentation/widgets/text_field_widget.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class TextFieldWidget extends StatefulWidget {
+  const TextFieldWidget({
+    super.key,
+    required this.controller,
+    required this.hintText,
+    required this.labelText,
+  });
+  final TextEditingController controller;
+  final String hintText, labelText;
+
+  @override
+  State<TextFieldWidget> createState() => _TextFieldWidgetState();
+}
+
+class _TextFieldWidgetState extends State<TextFieldWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      controller: widget.controller,
+      decoration: InputDecoration(
+        hintText: widget.hintText,
+        labelText: widget.labelText,
+      ),
+    );
+  }
+}